Move ClusterOperation to Cluster and rename it to Execute
